Fit Model

A function to fit the models in development through the full fitting pipeline based on synthetic data or observed data.

fit_model(model = "BaseLineModel", gen_data = TRUE, run_time = 74,
  time_scale = "year", plot_obs = TRUE, sample_priors = TRUE,
  prior_samples = 1000, nparticles = NULL, adapt_particles = FALSE,
  adapt_proposal = FALSE, min_acc = 0.05, max_acc = 0.4, fit = FALSE,
  posterior_samples = 1000, save_output = FALSE, nthreads = 4,
  verbose = TRUE, libbi_verbose = FALSE, browse = FALSE,
  const_pop = FALSE, no_age = FALSE, no_disease = FALSE)

Arguments

model
A character string containing the model name. Alternatively a model loaded via rbi::bi_model may be passed.

gen_data
Logical, defaults to /codeTRUE. Should data be synthesised using the model and priors.

run_time
Numeric, the number of years to run the model fitting and simulation for, defaults to 74 (i.e from 1931 until 2005).

time_scale
Character, defaults to "year". A monthly timescale can also be set with "month" .

plot_obs
Logical, defaults to TRUE. Should input data be plotted

sample_priors
Logical, defaults to FALSE. Should the model priors be sampled.

prior_samples
Numeric, the number of samples to take from the priors. Defaults to 1000.

nparticles
Numeric, the initial number of particles to use in the particle filters. Defaults to NULL, in which case the number of outputs is used as the initial particle number.

adapt_particles
Logical, defaults to FALSE. Should the number of particles be adapted.

adapt_proposal
Logical, defaults to TRUE. Should the proposal be adjusted to improve the acceptance rate.

min_acc
Numeric, defaults to 0.05. The minimum target acceptance rate.

max_acc
Numeric, defaults to 0.4. The maximum target acceptance rate.

fit
Logical, defaults to TRUE. Should the model be fitted with 1000 samples extracted.

posterior_samples
Numeric, the number of samples to take from the posterior estimated using pmcmc (requires fit = TRUE). Defaults to 1000.

save_output
Logical, defaults to FALSE. Should the model results be saved as a test dataset.

nthreads
Numeric, defaults to 4. The number of parallel jobs to run. The most efficient option is likely to be to match the number of cores available.

verbose
Logical, defaults to TRUE. Should progress messages and output be printed.

libbi_verbose
Logical, defaults to FALSE. Should libbi produce verbose progress and warnings messages.

browse
Logical, defaults to FALSE. Should the function be run in debug mode using browser.

const_pop
Logical, defaults to FALSE. Should a constant population be maintained using births equals deaths.

no_age
Logical, defaults to FALSE. Should ageing be disabled.

no_disease
Logical, defaults to FALSE. Should disease be removed from the model

Value

A LibBi model object based on the inputed test model.

Examples