Setup Observed Model Data

Performs data preprocessing required to prepare the observed data for the model.

setup_model_obs(years_of_age = NULL)

Arguments

years_of_age
Numeric, the years of age distributed cases to fit to. If not specified then no years are used.

Value

A named list of observed data required by the model.

Examples

## Code
setup_model_obs

#> function(years_of_age = NULL) {
#>   
#> 
#>   ## Extract historic Pulmonary TB cases
#>   historic_p_tb <- ModelTBBCGEngland::historic_cases %>%
#>     filter(year < 2000, year >= 1982) %>% 
#>     select(time = year, value = pulmonary) %>% 
#>     mutate(time = time - 1931)
#>   
#>   ## Extract age stratified UK born cases
#>   age_cases <- ModelTBBCGEngland::incidence %>% 
#>     filter(ukborn == "UK Born") %>% 
#>     select(-ukborn) %>% 
#>     group_by(year, age_group) %>% 
#>     summarise(value = sum(incidence, na.rm = T)) %>% 
#>     ungroup %>% 
#>     mutate(time = year - 1931) %>% 
#>     arrange(time, age_group) %>% 
#>     mutate(age = as.numeric(age_group) - 1) %>% 
#>     select(time, age, value) %>% 
#>     arrange(time, age)
#>   
#>   ## Extract UK born cases
#>   yearly_cases <- age_cases %>% 
#>     group_by(time) %>% 
#>     summarise(value = sum(value, na.rm = TRUE))
#>   
#>   ## Filter age cases
#>   if (is.null(years_of_age))  {
#>     age_cases <- age_cases %>% 
#>       filter(time %in% (years_of_age - 1931))
#>   }
#>   
#>   obs <- list(
#>     "YearlyHistPInc" = historic_p_tb,
#>     "YearlyAgeInc" = age_cases,
#>     "YearlyInc" = yearly_cases
#>   )
#>   
#>   return(obs)
#> }
#> <environment: namespace:ModelTBBCGEngland>

## Output
setup_model_obs()

#> $YearlyHistPInc
#> [90m# A tibble: 18 x 2[39m
#>     time value
#>    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m 1[39m    51  [4m5[24m466
#> [90m 2[39m    52  [4m4[24m880
#> [90m 3[39m    53  [4m4[24m535
#> [90m 4[39m    54  [4m4[24m293
#> [90m 5[39m    55  [4m4[24m345
#> [90m 6[39m    56  [4m3[24m640
#> [90m 7[39m    57  [4m3[24m675
#> [90m 8[39m    58  [4m3[24m816
#> [90m 9[39m    59  [4m3[24m618
#> [90m10[39m    60  [4m3[24m596
#> [90m11[39m    61  [4m3[24m816
#> [90m12[39m    62  [4m3[24m961
#> [90m13[39m    63  [4m3[24m694
#> [90m14[39m    64  [4m3[24m711
#> [90m15[39m    65  [4m3[24m690
#> [90m16[39m    66  [4m3[24m947
#> [90m17[39m    67  [4m3[24m926
#> [90m18[39m    68  [4m3[24m827
#> 
#> $YearlyAgeInc
#> [90m# A tibble: 0 x 3[39m
#> [90m# ... with 3 variables: time [3m[90m<dbl>[90m[23m, age [3m[90m<dbl>[90m[23m, value [3m[90m<int>[90m[23m[39m
#> 
#> $YearlyInc
#> [90m# A tibble: 16 x 2[39m
#>     time value
#>    [3m[90m<dbl>[39m[23m [3m[90m<int>[39m[23m
#> [90m 1[39m    69  [4m1[24m803
#> [90m 2[39m    70  [4m1[24m866
#> [90m 3[39m    71  [4m1[24m833
#> [90m 4[39m    72  [4m1[24m685
#> [90m 5[39m    73  [4m1[24m776
#> [90m 6[39m    74  [4m1[24m785
#> [90m 7[39m    75  [4m1[24m714
#> [90m 8[39m    76  [4m1[24m781
#> [90m 9[39m    77  [4m1[24m855
#> [90m10[39m    78  [4m1[24m900
#> [90m11[39m    79  [4m1[24m800
#> [90m12[39m    80  [4m1[24m940
#> [90m13[39m    81  [4m1[24m987
#> [90m14[39m    82  [4m1[24m818
#> [90m15[39m    83  [4m1[24m746
#> [90m16[39m    84  [4m1[24m537
#>