Fit Model

A function to fit the models in development through the full fitting pipeline based on synthetic data or observed data.

fit_model(model = "BaseLineModel", previous_model_path = NULL,
  gen_data = TRUE, run_time = 73, time_scale = "year",
  plot_obs = TRUE, sample_priors = TRUE, prior_samples = 1000,
  nparticles = NULL, adapt_particles = FALSE,
  adapt_part_samples = 1000, adapt_part_it = 10, min_particles = 4,
  max_particles = 16, proposal_param_block = NULL,
  proposal_initial_block = NULL, adapt_proposal = FALSE,
  adapt_prop_samples = 1000, adapt_prop_it = 10, adapt = "both",
  adapt_scale = 2, min_acc = 0.2, max_acc = 0.4, fit = FALSE,
  posterior_samples = 10000, thin = 10, burn_prop = 0,
  nthreads = parallel::detectCores(), verbose = TRUE,
  libbi_verbose = FALSE, fitting_verbose = TRUE, browse = FALSE,
  const_pop = FALSE, no_age = FALSE, no_disease = FALSE,
  save_output = FALSE, dir_path = NULL, dir_name = NULL,
  seed = 1234)

Arguments

model
A character string containing the model name. Alternatively a model loaded via rbi::bi_model may be passed.

previous_model_path
A character string giving the path to a previous LiBBi model (saved as an .rds). This will replace the default model and override initial settings.

gen_data
Logical, defaults to /codeTRUE. Should data be synthesised using the model and priors.

run_time
Numeric, the number of years to run the model fitting and simulation for, defaults to 74 (i.e from 1931 until 2005).

time_scale
Character, defaults to "year". A monthly timescale can also be set with "month" .

plot_obs
Logical, defaults to TRUE. Should input data be plotted

sample_priors
Logical, defaults to FALSE. Should the model priors be sampled.

prior_samples
Numeric, the number of samples to take from the priors. Defaults to 1000.

nparticles
Numeric, the initial number of particles to use in the particle filters. Defaults to NULL, in which case the number of outputs is used as the initial particle number.

adapt_particles
Logical, defaults to FALSE. Should the number of particles be adapted.

adapt_part_samples
Numeric, the number of samples to take from the priors when adapting the number of particles. Defaults to 1000.

adapt_part_it
Numeric, the number of iterations to use for adapting the number of particles. Defaults to 10.

min_particles
Numeric, defaults to 4. The starting number of particles to use for adpation.

max_particles
Numeric, defaults to 16. The maximum number of particles to use for apation.

proposal_param_block
A character string containing a LiBBi proposal parameter block. Defaults to NULL in which case the LiBBi defaults will be used.

proposal_initial_block
A character string containing a LiBBi proposal initial block. Defaults to NULL in which case the LiBBi defaults will be used.

adapt_proposal
Logical, defaults to TRUE. Should the proposal be adjusted to improve the acceptance rate.

adapt
Character string, defaults to "both". The type of adaption to use for the proposal see rbi.helpers::adapt_proposal for details.

adapt_scale
Numeric, defaults to 2. The scale to use to adapt the proposal.

min_acc
Numeric, defaults to 0.05. The minimum target acceptance rate.

max_acc
Numeric, defaults to 0.4. The maximum target acceptance rate.

fit
Logical, defaults to TRUE. Should the model be fitted with 1000 samples extracted.

posterior_samples
Numeric, the number of samples to take from the posterior estimated using pmcmc (requires fit = TRUE). Defaults to 1000.

thin
Numeric, the thinning interval to use between posterior samples. May reduce the correlation between samples and reduces memory.

burn_prop
Numeric (between 0 and 1). The proportion of the chain to discard as burn-in.

nthreads
Numeric, defaults to 4. The number of parallel jobs to run. The most efficient option is likely to be to match the number of cores available.

verbose
Logical, defaults to TRUE. Should progress messages and output be printed.

libbi_verbose
Logical, defaults to FALSE. Should libbi produce verbose progress and warnings messages.

fitting_verbose
Logical, defaults to FALSE. Should libbi produce verbose progress and warnings messages whilst fitting.

browse
Logical, defaults to FALSE. Should the function be run in debug mode using browser.

const_pop
Logical, defaults to FALSE. Should a constant population be maintained using births equals deaths.

no_age
Logical, defaults to FALSE. Should ageing be disabled.

no_disease
Logical, defaults to FALSE. Should disease be removed from the model

save_output
Logical, defaults to FALSE. Should the model results be saved as a test dataset.

seed
Numeric, the seed to use for random number generation.

adapt_part_samples
Numeric, the number of samples to take from the priors when adapting the proposal distribution. Defaults to 1000.

adapt_part_it
Numeric, defaults to 10. The number of iterations to use for adapting the proposal.

Value

A LibBi model object based on the inputed test model.

Examples

## Function code:
fit_model

#> function(model = "BaseLineModel", previous_model_path = NULL, gen_data = TRUE, 
#>                       run_time = 73, time_scale = "year", plot_obs = TRUE,
#>                       sample_priors = TRUE, prior_samples = 1000, nparticles = NULL,
#>                       adapt_particles = FALSE, adapt_part_samples = 1000, adapt_part_it = 10, 
#>                       min_particles = 4, max_particles = 16,
#>                       proposal_param_block = NULL, proposal_initial_block = NULL, 
#>                       adapt_proposal = FALSE, adapt_prop_samples = 1000, adapt_prop_it = 10, adapt = "both",
#>                       adapt_scale = 2, min_acc = 0.2, max_acc = 0.4,
#>                       fit = FALSE, posterior_samples = 10000, thin = 10, burn_prop = 0, 
#>                       nthreads = parallel::detectCores(), verbose = TRUE, libbi_verbose = FALSE, 
#>                       fitting_verbose = TRUE, browse = FALSE,
#>                       const_pop = FALSE, no_age = FALSE, no_disease = FALSE,
#>                       save_output = FALSE, dir_path = NULL, dir_name = NULL,
#>                       seed = 1234) {
#> 
#> 
#> # Util functions ----------------------------------------------------------
#> 
#>   ##Plot and save
#>   plot_obj <- function(obj, p_param, append_name = NULL, save = FALSE) {
#>     if (!is.null(p_param[[obj]]))
#>     {
#>       message("Plotting: ", obj)
#>       
#>       plot <- p_param[[obj]] + theme_minimal()
#>       
#>       print(plot)
#>       
#>       if (!is.null(append_name)) {
#>         obj <- paste0(append_name, "-", obj)
#>       }
#>       
#>       if (save) {
#>         ggsave(paste0(obj, ".png"), plot, path = plots_dir, dpi = 320)
#>       }
#> 
#>     }
#>   }
#> 
#> # Check parameters --------------------------------------------------------
#> 
#> if(burn_prop > 1 || burn_prop < 0) {
#>   stop("The burn in proportion must be less or equal to 1 and greater than or equal to 0.")
#> }  
#> 
#>   # Set up model directory --------------------------------------------------
#> if (save_output) {
#>   if (is.null(dir_path)) {
#>     dir_path <- "."
#>   }
#>   
#>   if (is.null(dir_name)) {
#>     dir_name <- "model-run"
#>   }
#>   
#>   ## Add time to filename
#>   dir_name <- paste0(dir_name, "-", str_replace(Sys.time(), " ", "_"))
#>   
#>   ## Construct path
#>   model_dir <- file.path(dir_path, dir_name)
#>   
#>   message("Model output: ", model_dir)
#>   
#>   ## Make the folder
#>   if (!dir.exists(model_dir)) {
#>     dir.create(model_dir)
#>     
#>     ## Make a sub folder for data
#>     data_dir <- file.path(model_dir, "data")
#>     
#>     dir.create(data_dir)
#>     
#>     ## Make a sub folder for plots
#>     plots_dir <- file.path(model_dir, "plots")
#>     
#>     dir.create(plots_dir)
#>     
#>     ## Make a sub folder for libbi output
#>     libbi_dir <- file.path(model_dir, "libbi")
#>     
#>     dir.create(libbi_dir)
#>   }
#>   
#>   ## Make the log file
#>   logs <- file.path(model_dir, "log.txt")
#>   con <- file(logs)
#>   
#>   ## Send Output to log
#>   sink(con, append = TRUE)
#>   sink(con, type = "message", append = TRUE)
#> }
#>   
#>   # Set the time scale for the model ----------------------------------------
#> 
#> 
#>   if (time_scale == "year") {
#>     time_scale_numeric <- 1
#>   }else if (time_scale == "month") {
#>     time_scale_numeric <- 12
#>   }else{
#>     stop("Only a yearly (year) or monthly (month) timescale have been implemented.")
#>   }
#>   
#> 
#>   # Set the number of particles ---------------------------------------------
#> 
#>   if (is.null(nparticles)) {
#>     nparticles <- run_time * time_scale_numeric
#>   }  
#>   
#>   # Load the model ----------------------------------------------------------
#>   
#>   if (is.character(model)) {
#>     model_file <- system.file(package="ModelTBBCGEngland", paste0("bi/", model, ".bi"))
#>     
#>     tb_model_raw <- bi_model(model_file)
#>     
#>   }else{
#>     tb_model_raw <- model
#>   }
#>   
#>   
#> 
#> # Specify model setup conditions ------------------------------------------
#> 
#>   if (time_scale == "year") {
#>     tb_model_raw <- fix(tb_model_raw, ScaleTime = 1 / 12)
#>   }else if (time_scale == "month") {
#>     tb_model_raw <- fix(tb_model_raw, ScaleTime = 1)
#>   }else{
#>     stop("Only a yearly (year) or monthly (month) timescale have been implemented.")
#>   }
#>   
#>   if (const_pop) {
#>     tb_model_raw <- fix(tb_model_raw, const_pop = 1)
#>   }
#>   
#>   if (no_age) {
#>     tb_model_raw <- fix(tb_model_raw, no_age = 1)
#>   }
#>   
#>   if (no_disease) {
#>     tb_model_raw <- fix(tb_model_raw, no_disease = 1)
#>   }
#>   
#> 
#> # Add the proposal block to the model -------------------------------------
#>  if (!is.null(proposal_param_block)) {
#>    tb_model_raw <- add_block(tb_model_raw, "proposal_parameter", proposal_param_block)
#>  }
#>   
#>   
#>   if (!is.null(proposal_initial_block)) {
#>     tb_model_raw <- add_block(tb_model_raw, "proposal_initial", proposal_initial_block)
#>   }
#> 
#> 
#>   # Allow for interactive debugging -----------------------------------------
#>   
#>   if (browse) {
#>     browser() 
#>   }
#> 
#> 
#> # Set up model input ------------------------------------------------------
#> 
#> ## See ?setup_model_input or details 
#> input <- setup_model_input(run_time, time_scale_numeric)
#> 
#> ## Save formated data
#> if (save_output) {
#>   saveRDS(input, file.path(data_dir, "input.rds"))
#> }
#> 
#> 
#> # Set up abs data ---------------------------------------------------------
#> 
#> ## See ?setup_model_obs for details
#> obs <- setup_model_obs()
#> 
#> # Reset obs and input if running with test SIR Model ----------------------
#>   if (model == "SIR") {
#>     obs <- NULL
#>     input <- NULL
#>   }
#>   
#> # Generate data from the model --------------------------------------------
#>   if (gen_data) {
#>     
#>     if (verbose) {
#>       message("Generating data from model")
#>     }
#>     
#>     tb_data <- bi_generate_dataset(tb_model_raw, end_time = run_time * time_scale_numeric, 
#>                                    input = input, obs = obs, 
#>                                    noutputs = floor(run_time / 7),
#>                                    seed = seed, verbose = libbi_verbose)
#>     
#>     if (verbose) {
#>       message("Summary of generated model data")
#>       print(tb_data)
#>     }
#>     
#>     
#>     obs <- bi_read(tb_data, type = "obs")
#>     
#>     if (verbose) {
#>       message("Variables in the generated data")
#>       print(names(obs))
#>     }
#>     
#>   }
#>   
#>   ## Save formated data
#>   if (save_output) {
#>     saveRDS(obs, file.path(data_dir, "obs.rds"))
#>   }
#>   
#>   
#>   # Plot incidence generated by the model -----------------------------------
#>   if (plot_obs) {
#>     
#>     time <- NULL; value <- NULL;
#>     
#>     message("Plot historic number of pulmonary cases detected each year:")
#>     p_hist_cases <- obs$YearlyHistPInc %>% 
#>       dplyr::filter(time > 0) %>% 
#>       ggplot(aes(x = time, y = value)) +
#>       geom_point(size = 1.2) +
#>       geom_line(size = 1.1, alpha = 0.6) +
#>       theme_minimal() +
#>       labs(x = "Time",
#>            y = "Yearly notified cases")
#>     
#>     print(p_hist_cases)
#>     
#>     if (save_output) {
#>       ggsave("obs-hist-pul-cases.png", path = plots_dir, dpi = 320)
#>     }
#>     
#>     message("Plot number of cases detected each year:")
#>     p_cases <- obs$YearlyInc %>% 
#>       dplyr::filter(time > 0) %>% 
#>       ggplot(aes(x = time, y = value)) +
#>       geom_point(size = 1.2) +
#>       geom_line(size = 1.1, alpha = 0.6) +
#>       theme_minimal() +
#>       labs(x = "Time",
#>            y = "Yearly notified cases")
#>     
#>     print(p_cases)
#>     if (save_output) {    
#>       ggsave("obs-cases.png", path = plots_dir, dpi = 320)
#>     }
#>     
#>   message("Plot number of cases detected each year, stratified by age group:")
#>     p_age_cases <- obs$YearlyAgeInc %>% 
#>       dplyr::filter(time > 0) %>% 
#>       ggplot(aes(x = time, y = value)) +
#>       geom_point(size = 1.2) +
#>       geom_line(size = 1.1, alpha = 0.6) +
#>       theme_minimal() +
#>       labs(x = "Time",
#>            y = "Yearly notified cases") +
#>       facet_wrap(~age, scales = "free_y")
#>     
#>     print(p_age_cases)
#>     
#>     if (save_output) {
#>       ggsave("obs-age-cases.png", path = plots_dir, dpi = 320)
#>     }
#> 
#>   }
#> 
#>   
#>   # Load the model ----------------------------------------------------------
#>   
#>   if (verbose) {
#>     message("Load the model as a compiled libbi object")
#>   }
#> 
#>   tb_model <- libbi(tb_model_raw, 
#>                     input = input, 
#>                     obs = obs,
#>                     end_time = run_time * time_scale_numeric, 
#>                     noutputs = run_time * time_scale_numeric,
#>                     nparticles = nparticles, nthreads = nthreads, 
#>                     verbose = libbi_verbose,
#>                     seed = seed)
#>   
#>   if (!is.null(previous_model_path)) {
#>     message("Replacing the default liBBi model with a previously run model - this may not have the same settings as the current run.")
#>     tb_model <- read_libbi(previous_model_path)
#>     
#>     if (verbose) {
#>       message("Previous Model: ")
#>       print(tb_model) 
#>       print(tb_model$model)
#>     }
#>   }
#>   # Sample from priors ------------------------------------------------------
#> if (sample_priors) {
#>   if (verbose) {
#>     message("Sample priors")
#>   }
#>   
#>   priors <- sample(tb_model, target = "prior", nsamples = prior_samples)
#>   
#>   if (verbose) {
#>     message("Summary of prior sampling")
#>     print(priors)
#>     message("Load priors")
#>   }
#>   
#>   
#>   if (verbose) {
#>     message("Plot priors")
#>   
#>     p_prior <- plot(priors , plot = FALSE)
#>     
#>     objects <- c("states", "densities", "traces", "correlations", "noises", "likelihoods")
#>     
#>     map(objects, ~ plot_obj(., p_prior, append_name = "prior", save = save_output))
#>       
#>     if (save_output) {
#>       saveRDS(p_prior$data, file.path(data_dir, "prior-params.rds"))
#>     }
#> 
#>   }
#> }  
#> 
#>   # Adapting the number of particles ----------------------------------------
#>   
#>   if (adapt_particles) {
#>     if (verbose) {
#>       message("Adapting particles starting with ", min_particles, " up to a maximum of ", max_particles)
#>     }
#>     
#>     ## Add all variables as outputs to the model.
#>     lim_out_model <- tb_model$model
#>     tb_model$model <- everything_from_model(tb_model$model)
#> 
#>     adapt_mutli_particles <- function(iteration, libbi) {
#>       if (verbose) {
#>         message("Adapting particles iteration: ", iteration)
#>         message("Initial sampling using the prior as the proposal.")
#>       }
#>       
#>       tb_model <- tb_model %>% 
#>         sample(proposal = "prior", nsamples = adapt_part_samples, verbose = libbi_verbose,
#>                options = list(with="transform-initial-to-param"), seed = iteration + seed)
#>       
#>       if (verbose) {
#>         message("Starting particle adaption")
#>       }
#>       
#>       tb_model <- tb_model %>% 
#>         adapt_particles(min = min_particles, max = max_particles, 
#>                         quiet = !verbose,
#>                         verbose = libbi_verbose)
#>       
#>       particles <- tb_model$options$nparticles
#>       return(particles)
#>     }
#>     
#>     particles <- map_dbl(1:adapt_part_it, ~  adapt_mutli_particles(., tb_model))
#> 
#>     
#>     if (save_output) {
#>       saveRDS(particles, file.path(libbi_dir, "particles.rds"))
#>     }
#>     
#>     med_particles <- particles %>% 
#>       median %>% 
#>       ceiling
#>     
#>     if (verbose) {
#>       message("Choosing ", med_particles, "based on ", adapt_part_it, " iterations each using ", adapt_part_samples, ".")
#>       message("The maximum number of particles selected was ", max(particles), " with a minimum of", min(particles))
#>       message("The mean number of particles selected was ", mean(particles), " with a standard deviation of ", sd(particles))
#>     }
#>     
#>  
#>     tb_model$model <- lim_out_model
#>     tb_model$options$nparticles <- med_particles
#>     
#>   }
#>   
#>   
#>   # Adapting the proposal ---------------------------------------------------
#>   
#>   if (adapt_proposal) {
#>     if (verbose) {
#>       message("Adapting proposal with a min acceptance of ", min_acc, " and a maximum acceptance of ", max_acc)
#>       message("Running for ", adapt_prop_it, " iterations with ", adapt_prop_samples, " samples each time.")
#>     }
#>     
#>     tb_model <- tb_model %>% 
#>       sample(proposal = "model", nsamples = adapt_prop_samples, verbose = libbi_verbose) %>% 
#>       adapt_proposal(min = min_acc, max = max_acc,
#>                      max_iter = adapt_prop_it, adapt = adapt, 
#>                      scale = adapt_scale, truncate = TRUE, 
#>                      quiet = !verbose)
#>     
#>     prop_param_block <- get_block(tb_model$model, "proposal_parameter")
#>     prop_initial_block <- get_block(tb_model$model, "proposal_initial")
#>     
#>     if (verbose) {
#>       message("Adapted proposal distribution")
#>     
#>       print(prop_param_block)
#>       print(prop_initial_block)
#>     }
#>     
#>     if (save_output) {
#>       saveRDS(prop_param_block, file.path(libbi_dir, "proposal-param-block.rds"))
#>       saveRDS(prop_initial_block, file.path(libbi_dir, "proposal-initial-block.rds"))
#>     }
#>   }
#>   
#>   
#>   # Model fitting ------------------------------------------------------
#>   
#>   
#>   if (fit) {
#>     if (verbose) {
#>       message("Fitting using PMCMC")
#>     }
#>     
#>     
#>     tb_model <- tb_model %>% 
#>       sample(target = "posterior",
#>              proposal = "model", sample_obs = TRUE, 
#>              nsamples = posterior_samples,
#>              thin = thin,
#>              verbose = fitting_verbose)
#>     
#>     
#>     if (verbose) {
#>       message("Summary of fitted model")
#>       print(tb_model)
#>     }
#>     
#>     if (save_output) {
#>       save_libbi(tb_model, file.path(libbi_dir, "posterior.rds"))
#>     }
#>     
#> 
#> # Plot posterior ----------------------------------------------------------
#> 
#>     if (verbose) {
#>       message("Plot posterior")
#>       
#>       p_posterior <- plot(tb_model, plot = FALSE)
#>       
#>       objects <- c("states", "densities", "traces", "correlations", "noises", "likelihoods")
#>       
#>       map(objects, ~ plot_obj(., p_posterior, append_name = "posterior", save = save_output))
#>       
#>       if (save_output) {
#>         saveRDS(p_posterior$data, file.path(data_dir, "posterior-params.rds"))
#>       }
#>       
#>     }
#>     
#>     if (save_output) {
#>       save_libbi(tb_model, file.path(libbi_dir, "posterior"))
#>     }
#>     
#> 
#> # Evaluate run ------------------------------------------------------------
#>     
#>     burn <- floor(posterior_samples / thin * burn_prop)
#>     
#>     dic <- DIC(tb_model, burn = burn)
#>     
#>     message("Model DIC: ", dic)
#>     
#>     if (save_output) {
#>       saveRDS(dic, file = file.path(libbi_dir, "dic.rds"))
#>     }
#> 
#>   }
#>   
#> 
#>   if (save_output) {
#>     sink(file = NULL) 
#>     sink(file = NULL, type = "message")
#>   }
#>   
#>   if (!fit && !adapt_particles && !adapt_proposal && sample_priors) {
#>     tb_model <- priors
#>   }
#>   
#>   return(tb_model)
#> }
#> <environment: namespace:ModelTBBCGEngland>