Setup Input Model Data

Performs data preprocessing required to prepare the input data for the model.

setup_model_input(run_time = NULL, time_scale_numeric = 1)

Arguments

run_time
Numeric, the run time of the model

time_scale_numeric
Numeric the time scale to use (with 1 being a year, 12 a month etc.). Defaults to 1.

Value

A named list of data inputs required by the model.

Examples

## Code
setup_model_input

#> function(run_time = NULL, time_scale_numeric = 1) {
#>   
#>   ## Set up initial population distribution
#>   pop_dist <- england_demographics %>% 
#>     filter(CoB == "UK born") %>% 
#>     group_by(year) %>% 
#>     mutate(age = 0:(n() - 1)) %>% 
#>     group_by(age) %>% 
#>     summarise(value = mean(proportion_age_by_year)) %>% 
#>     select(age, value)
#>   
#>   ## Set up births scaling for time horizon
#>   t_births <- births %>% 
#>     filter(year >= 1931) %>% 
#>     mutate(time = year - 1931) %>% 
#>     mutate(time_n = map(time, ~ tibble(time_n = time_scale_numeric * . + 0:(time_scale_numeric - 1)))) %>% 
#>     unnest() %>% 
#>     mutate(value = births / time_scale_numeric) %>% 
#>     select(time = time_n, value) %>% 
#>     filter(time <= run_time * time_scale_numeric)
#>   
#>   ## Set up expected lifespan
#>   exp_life_span <- mortality_rates %>% 
#>     mutate(time = year - 1931,
#>            value = exp_life_span * time_scale_numeric) %>% 
#>     group_by(time) %>% 
#>     mutate(age = 0:(n() - 1)) %>% 
#>     ungroup %>% 
#>     mutate(time_n = map(time, ~ tibble(time_n = time_scale_numeric * . + 0:(time_scale_numeric - 1)))) %>% 
#>     unnest() %>% 
#>     select(time = time_n, age, value) %>% 
#>     filter(time <= run_time * time_scale_numeric)
#>   
#>   ## Set up Polymod contacts 
#>   polymod <- contact %>% 
#>     arrange(age_x, age_y) %>% 
#>     group_by(age_x) %>% 
#>     mutate(age2 = 0:(n() - 1)) %>% 
#>     group_by(age_y) %>% 
#>     mutate(age = 0:(n() - 1)) %>% 
#>     ungroup %>% 
#>     mutate_at(.vars = vars(mean, sd),
#>               ~ . / time_scale_numeric) %>% 
#>     select(age, age2, mean, sd)
#>   
#>   ## Mean contacts
#>   polymod_mean <- polymod %>% 
#>     select(age, age2, value = mean)
#>   
#>   ## SD of contacts
#>   polymod_sd <- polymod %>% 
#>     select(age, age2, value = sd)
#>   
#>   
#>   ## Extact non-UK born pulmonary cases - estimate previous cases in the model
#>   nonukborn_p_cases <- incidence %>% 
#>     filter(ukborn == "Non-UK Born",
#>            pulmextrapulm == "Pulmonary, with or without EP") %>% 
#>     select(-ukborn, -pulmextrapulm, -type, -policy_change) %>% 
#>     mutate(time = year - 1931) %>% 
#>     arrange(time, age_group) %>% 
#>     mutate(age = as.numeric(age_group) - 1) %>% 
#>     select(time, age, incidence) %>% 
#>     count(time, age, wt = incidence) %>% 
#>     rename(value = n) %>% 
#>     mutate(time_n = map(time, ~ tibble(time_n = time_scale_numeric * . + 0:(time_scale_numeric - 1)))) %>% 
#>     unnest() %>% 
#>     mutate(value = value / time_scale_numeric) %>% 
#>     select(time = time_n, age, value)
#>   
#>   ## Estimated future non UK born cases using a poisson regression model adjusting for time and age.
#>     nonukborn_p_cases <-  nonukborn_p_cases  %>% 
#>     dplyr::filter(time >= (2010 - 1931)) %>% 
#>     nest() %>% 
#>     mutate(model = map(data, ~ glm(value ~ time + factor(age),, family = poisson, data = .))) %>% 
#>     mutate(new_data = map(data,  ~ expand.grid(age = min(.$age):max(.$age), time = c(max(.$time) + 1):(2100 - 1931)))) %>% 
#>     mutate(pred_cases = map2(model, new_data, ~ tibble(value = predict(.x, .y, type = "response")))) %>% 
#>     mutate(pred_data = map2(new_data, pred_cases, ~ bind_cols(.x, .y))) %>% 
#>     mutate(all_data = map(pred_data, ~ bind_rows(nonukborn_p_cases, .))) %>% 
#>     select(all_data) %>% 
#>     unnest(all_data)
#>   
#>   
#>   ## Non UK born cases in 2000 - used to estimate historic non UK born cases
#>   NUKCases2000 <- nonukborn_p_cases %>% 
#>     filter(time == time_scale_numeric * (2000 - 1931)) %>% 
#>     select(-time)
#>   
#>   
#>   
#>   input <- list(
#>     "pop_dist" = pop_dist,
#>     "births_input" = t_births,
#>     "exp_life_span" = exp_life_span,
#>     "polymod" = polymod_mean,
#>     "polymod_sd" = polymod_sd,
#>     "NonUKBornPCases" = nonukborn_p_cases,
#>     "NUKCases2000" = NUKCases2000
#>   )  
#>   
#>   return(input)
#> }
#> <environment: namespace:ModelTBBCGEngland>

## Output
setup_model_input(run_time = 80, time_scale_numeric = 1)

#> $pop_dist
#> [90m# A tibble: 15 x 2[39m
#>      age  value
#>    [3m[90m<int>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m 1[39m     0 0.068[4m2[24m
#> [90m 2[39m     1 0.065[4m1[24m
#> [90m 3[39m     2 0.065[4m1[24m
#> [90m 4[39m     3 0.065[4m0[24m
#> [90m 5[39m     4 0.063[4m5[24m
#> [90m 6[39m     5 0.060[4m9[24m
#> [90m 7[39m     6 0.062[4m7[24m
#> [90m 8[39m     7 0.067[4m3[24m
#> [90m 9[39m     8 0.070[4m3[24m
#> [90m10[39m     9 0.068[4m4[24m
#> [90m11[39m    10 0.065[4m2[24m
#> [90m12[39m    11 0.061[4m3[24m
#> [90m13[39m    12 0.055[4m9[24m
#> [90m14[39m    13 0.049[4m6[24m
#> [90m15[39m    14 0.112 
#> 
#> $births_input
#> [90m# A tibble: 81 x 2[39m
#>     time  value
#>    [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m 1[39m     0 [4m5[24m[4m8[24m[4m9[24m751
#> [90m 2[39m     1 [4m5[24m[4m7[24m[4m3[24m099
#> [90m 3[39m     2 [4m5[24m[4m4[24m[4m0[24m989
#> [90m 4[39m     3 [4m5[24m[4m5[24m[4m7[24m686
#> [90m 5[39m     4 [4m5[24m[4m5[24m[4m9[24m569
#> [90m 6[39m     5 [4m5[24m[4m6[24m[4m7[24m383
#> [90m 7[39m     6 [4m5[24m[4m7[24m[4m3[24m382
#> [90m 8[39m     7 [4m5[24m[4m8[24m[4m3[24m579
#> [90m 9[39m     8 [4m5[24m[4m8[24m[4m1[24m950
#> [90m10[39m     9 [4m5[24m[4m6[24m[4m7[24m710
#> [90m# ... with 71 more rows[39m
#> 
#> $exp_life_span
#> [90m# A tibble: 1,215 x 3[39m
#>     time   age value
#>    [3m[90m<dbl>[39m[23m [3m[90m<int>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m 1[39m     0     0  75.3
#> [90m 2[39m     0     1 777. 
#> [90m 3[39m     0     2 955. 
#> [90m 4[39m     0     3 563. 
#> [90m 5[39m     0     4 793. 
#> [90m 6[39m     0     5 988. 
#> [90m 7[39m     0     6 909. 
#> [90m 8[39m     0     7 635. 
#> [90m 9[39m     0     8 317. 
#> [90m10[39m     0     9 128. 
#> [90m# ... with 1,205 more rows[39m
#> 
#> $polymod
#> [90m# A tibble: 225 x 3[39m
#>      age  age2 value
#>    [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m 1[39m     0     0  700.
#> [90m 2[39m     0     1  293.
#> [90m 3[39m     0     2  173.
#> [90m 4[39m     0     3  110.
#> [90m 5[39m     0     4  177.
#> [90m 6[39m     0     5  289.
#> [90m 7[39m     0     6  322.
#> [90m 8[39m     0     7  395.
#> [90m 9[39m     0     8  158.
#> [90m10[39m     0     9  101.
#> [90m# ... with 215 more rows[39m
#> 
#> $polymod_sd
#> [90m# A tibble: 225 x 3[39m
#>      age  age2 value
#>    [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m 1[39m     0     0 119. 
#> [90m 2[39m     0     1  32.5
#> [90m 3[39m     0     2  28.5
#> [90m 4[39m     0     3  20.8
#> [90m 5[39m     0     4  34.7
#> [90m 6[39m     0     5  34.8
#> [90m 7[39m     0     6  37.8
#> [90m 8[39m     0     7  68.7
#> [90m 9[39m     0     8  32.3
#> [90m10[39m     0     9  19.7
#> [90m# ... with 215 more rows[39m
#> 
#> $NonUKBornPCases
#> [90m# A tibble: 1,515 x 3[39m
#>     time   age value
#>    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m 1[39m    69     0    15
#> [90m 2[39m    69     1     6
#> [90m 3[39m    69     2    32
#> [90m 4[39m    69     3    92
#> [90m 5[39m    69     4   216
#> [90m 6[39m    69     5   260
#> [90m 7[39m    69     6   245
#> [90m 8[39m    69     7   131
#> [90m 9[39m    69     8   110
#> [90m10[39m    69     9    84
#> [90m# ... with 1,505 more rows[39m
#> 
#> $NUKCases2000
#> [90m# A tibble: 15 x 2[39m
#>      age value
#>    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m 1[39m     0    15
#> [90m 2[39m     1     6
#> [90m 3[39m     2    32
#> [90m 4[39m     3    92
#> [90m 5[39m     4   216
#> [90m 6[39m     5   260
#> [90m 7[39m     6   245
#> [90m 8[39m     7   131
#> [90m 9[39m     8   110
#> [90m10[39m     9    84
#> [90m11[39m    10    79
#> [90m12[39m    11    65
#> [90m13[39m    12    98
#> [90m14[39m    13    88
#> [90m15[39m    14   154
#>