Setup Observed Model Data

Performs data preprocessing required to prepare the observed data for the model.

setup_model_obs()

Value

A named list of observed data required by the model.

Examples

## Code
setup_model_obs

#> function() {
#>   
#>   ## Extract historic Pulmonary TB cases
#>   historic_p_tb <- ModelTBBCGEngland::historic_cases %>%
#>     filter(year < 2000, year >= 1990) %>% 
#>     select(time = year, value = pulmonary) %>% 
#>     mutate(time = time - 1931)
#>   
#>   ## Extract age stratified UK born cases
#>   age_cases <- ModelTBBCGEngland::incidence %>% 
#>     filter(ukborn == "UK Born") %>% 
#>     select(-ukborn) %>% 
#>     group_by(year, age_group) %>% 
#>     summarise(value = sum(incidence, na.rm = T)) %>% 
#>     ungroup %>% 
#>     mutate(time = year - 1931) %>% 
#>     arrange(time, age_group) %>% 
#>     mutate(age = as.numeric(age_group) - 1) %>% 
#>     select(time, age, value) %>% 
#>     arrange(time, age)
#>   
#>   ## Extract UK born cases
#>   yearly_cases <- age_cases %>% 
#>     group_by(time) %>% 
#>     summarise(value = sum(value, na.rm = TRUE))
#>   
#>   obs <- list(
#>     "YearlyHistPInc" = historic_p_tb,
#>     "YearlyAgeInc" = age_cases,
#>     "YearlyInc" = yearly_cases
#>   )
#>   
#>   return(obs)
#> }
#> <environment: namespace:ModelTBBCGEngland>

## Output
setup_model_obs()

#> $YearlyHistPInc
#> # A tibble: 10 x 2
#>     time value
#>    <dbl> <dbl>
#>  1    59  3618
#>  2    60  3596
#>  3    61  3816
#>  4    62  3961
#>  5    63  3694
#>  6    64  3711
#>  7    65  3690
#>  8    66  3947
#>  9    67  3926
#> 10    68  3827
#> 
#> $YearlyAgeInc
#> # A tibble: 240 x 3
#>     time   age value
#>    <dbl> <dbl> <int>
#>  1    69     0    75
#>  2    69     1    59
#>  3    69     2    75
#>  4    69     3   112
#>  5    69     4   133
#>  6    69     5   116
#>  7    69     6   147
#>  8    69     7    99
#>  9    69     8    83
#> 10    69     9   105
#> # ... with 230 more rows
#> 
#> $YearlyInc
#> # A tibble: 16 x 2
#>     time value
#>    <dbl> <int>
#>  1    69  1803
#>  2    70  1866
#>  3    71  1833
#>  4    72  1685
#>  5    73  1776
#>  6    74  1785
#>  7    75  1714
#>  8    76  1781
#>  9    77  1855
#> 10    78  1900
#> 11    79  1800
#> 12    80  1940
#> 13    81  1987
#> 14    82  1818
#> 15    83  1746
#> 16    84  1537
#>